+++ src/Core/CustomersManagement.Application/Features/Tours/Queries/GetTourDetails/TourDetailsDto.cs
@@ -6,8 +6,8 @@ public class TourDetailsDto
     public string Name { get; set; }
     public string TimeStart { get; set; }
     public string TimeEnd { get; set; }
-    public int EntireCost { get; set; }
-    public int AdvancePaymentCost { get; set; }
+    public double EntireCost { get; set; }
+    public double AdvancePaymentCost { get; set; }
     public string EntireAmountPaymentDeadline { get; set; }
     public string AdvancePaymentDeadline { get; set; }
     public string Status { get; set; }
